Přesun metody pro generování akčních tlačítek pro správu předmětů v obchodu do jiné třídy.
