+++ src/main/java/cz/stechy/drd/controller/bestiary/BestiaryHelper.java
@@ -110,4 +110,6 @@ public final class BestiaryHelper {
             .build();
     }
 
+
+
 }
