+++ src/main/java/cz/stechy/drd/controller/shop/ShopBackpackController.java
@@ -112,7 +112,7 @@ public class ShopBackpackController implements Initializable, ShopItemController
     @Override
     public void setShoppingCart(IShoppingCart shoppingCart, OnUploadItem uploadHandler,
         OnDownloadItem downloadHandler, OnDeleteItem deleteHandler) {
-        columnAction.setCellFactory(param -> CellUtils
+        columnAction.setCellFactory(param -> ShopHelper
             .forActionButtons(shoppingCart::addItem, shoppingCart::removeItem, uploadHandler,
                 downloadHandler, deleteHandler, user, resources));
     }
