+++ src/main/java/cz/stechy/drd/controller/shop/ShopGeneralController.java
@@ -109,7 +109,7 @@ public class ShopGeneralController implements Initializable, ShopItemController 
     @Override
     public void setShoppingCart(IShoppingCart shoppingCart, OnUploadItem uploadHandler,
         OnDownloadItem downloadHandler, OnDeleteItem deleteHandler) {
-        columnAction.setCellFactory(param -> CellUtils
+        columnAction.setCellFactory(param -> ShopHelper
             .forActionButtons(shoppingCart::addItem, shoppingCart::removeItem, uploadHandler,
                 downloadHandler, deleteHandler, user, resources));
     }
