+++ src/main/java/cz/stechy/drd/controller/shop/ShopWeaponMeleController.java
@@ -137,7 +137,7 @@ public class ShopWeaponMeleController implements Initializable, ShopItemControll
     @Override
     public void setShoppingCart(IShoppingCart shoppingCart, OnUploadItem uploadHandler,
         OnDownloadItem downloadHandler, OnDeleteItem deleteHandler) {
-        columnAction.setCellFactory(param -> CellUtils
+        columnAction.setCellFactory(param -> ShopHelper
             .forActionButtons(shoppingCart::addItem, shoppingCart::removeItem, uploadHandler,
                 downloadHandler, deleteHandler, user, resources));
     }
